# swagger-doc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tokens"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные запроса"
          },
          "403": {
            "description": "Email уже используется"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signin": {
      "post": {
        "operationId": "AuthController_signin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SigninDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/quiz/create": {
      "post": {
        "operationId": "QuizController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Quiz"
        ]
      }
    },
...
    "/quiz/{id}": {
      "get": {
        "operationId": "QuizController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Quiz"
        ]
      },
      "patch": {
        "operationId": "QuizController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Quiz"
        ]
      },
      "delete": {
        "operationId": "QuizController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Quiz"
        ]
      }
    },
    "/quiz/findAllByQuizId/{id}": {
      "get": {
        "operationId": "QuizController_findAllByQuizId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Quiz"
        ]
      }
    },
    "/question/{quizId}": {
      "post": {
        "operationId": "QuestionController_create",
        "parameters": [
          {
            "name": "quizId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Question"
        ]
      }
    },
    "/option/{questionId}": {
      "post": {
        "operationId": "OptionController_createOption",
        "parameters": [
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Option"
        ]
      },
      "get": {
        "operationId": "OptionController_findAllByQuestion",
        "parameters": [
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Option"
        ]
      }
    },
    "/option/item/{id}": {
      "get": {
        "operationId": "OptionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Option"
        ]
      }
    },
    "/option/{id}": {
      "patch": {
        "operationId": "OptionController_updateOption",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Option"
        ]
      },
      "delete": {
        "operationId": "OptionController_removeOption",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Option"
        ]
      }
    },
    "/user-quiz-result/start": {
      "post": {
        "operationId": "UserQuizResultController_startQuiz",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartQuizDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "UserQuizResult"
        ]
      }
    },
    "/user-quiz-result/finish/{resultId}": {
      "patch": {
        "operationId": "UserQuizResultController_finishQuiz",
        "parameters": [
          {
            "name": "resultId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserQuizResult"
        ]
      }
    },
    "/user-quiz-result/{id}": {
      "get": {
        "operationId": "UserQuizResultController_getResultById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserQuizResult"
        ]
      }
    },
...
    "/user-quiz-result/byQuiz/{quizId}": {
      "get": {
        "operationId": "UserQuizResultController_getResultsByQuiz",
        "parameters": [
          {
            "name": "quizId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserQuizResult"
        ]
      }
    },
    "/user-question-answer": {
      "post": {
        "operationId": "UserQuestionAnswerController_submitAnswer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserQuestionAnswerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "UserQuestionAnswer"
        ]
      }
    },
    "/user-question-answer/{answerId}": {
      "patch": {
        "operationId": "UserQuestionAnswerController_updateAnswer",
        "parameters": [
          {
            "name": "answerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserQuestionAnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserQuestionAnswer"
        ]
      }
    }
  },
  "info": {
    "title": "Your API",
    "description": "API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "username"
        ]
      },
      "Tokens": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT Access Token"
          },
          "refresh_token": {
            "type": "string",
            "description": "JWT Refresh Token"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ]
      },
      "SigninDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateQuizDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the quiz"
          },
          "description": {
            "type": "string",
            "description": "The description of the quiz"
          },
          "isPublic": {
            "type": "boolean",
            "description": "Indicates if the quiz is public"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateQuizDto": {
        "type": "object",
        "properties": {}
      },
      "CreateQuestionDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "The text of the question"
          },
          "type": {
            "type": "string",
            "description": "The type of the question",
            "enum": [
              "single_choice",
              "multiple_choice",
              "text",
              "ordering",
              "matching"
            ]
          },
          "score": {
            "type": "number",
            "description": "The score of the question",
            "default": 1
          }
        },
        "required": [
          "text",
          "type"
        ]
      },
      "CreateOptionDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          }
        },
        "required": [
          "text"
        ]
      },
      "UpdateOptionDto": {
        "type": "object",
        "properties": {}
      },
      "StartQuizDto": {
        "type": "object",
        "properties": {}
      },
      "CreateUserQuestionAnswerDto": {
        "type": "object",
        "properties": {
          "userQuizResultId": {
            "type": "string"
          },
          "questionId": {
            "type": "string"
          },
          "chosenOptionIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "textAnswer": {
            "type": "string"
          },
          "score": {
            "type": "number"
          }
        },
        "required": [
          "userQuizResultId",
          "questionId"
        ]
      },
      "UpdateUserQuestionAnswerDto": {
        "type": "object",
        "properties": {
          "userQuizResultId": {
            "type": "string"
          },
          "questionId": {
            "type": "string"
          },
          "chosenOptionIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "textAnswer": {
            "type": "string"
          },
          "score": {
            "type": "number"
          }
        }
      }
    }
  }
}